feat: disable co-pilot
My productivity had been profound and i feel at peace from not getting over-reliant on
copilot. By the way, am no  anti-AI, in fact would be great when am
working but defintely not for personal projects or work.

Those are meant for growth, to relish in the pain and take home the
experience. Forge them into my mind like a true war brat.